(#2) Allow configurable replacements for / and .
Prior to this commit, we were sending the URI portions of the metric
names off to statsd untransformed, which meant the metrics using
contained a series of `/`s in them. This was fine for statsd, but
graphite does not care for `/` in the metric name. Additionally, we
sometimes have legitimate periods/full stops `.` as part of the URI, but
this should not translate into a sub-portion of the metric; i.e.
`/api/v1/users/jeff.weiss` becoming `api.v1.users.jeff.weiss` is
confusing, therefore we first transform any period/full stop `.` into
`_`, which would yield the clearer `api.v1.users.jeff_weiss` for the
given example.
We also allow the users to configure a replacement for `.` other than
`_` and a replacement for `/` other than `.`.

Close #2.