(#3) Enable configurable metrics
Prior to this commit, the payloads sent by Plug.Statsd were static in
shape; the user had no way (other than forking and changing code) to
alter which aspects of a request/response were part of a statsd metric
and in which order. This commit allows for the metrics to be defined in
the mix config and provides a path for the user to select which elements
they wish to be part of their statsd metric name, in which order,
include any custom text if desired, and set the sampling rate for that
metric.

Close #3.