ensure backend is available
Prior to this commit, we didn't actually attempt to ensure that the
backend application had be started prior to us attempting to make calls
to it. This commit attempts to start the specified backend as soon as
the plug is initialized so that if it is not available, the plug will
immediately fail.